images/gradio_display/requirements.txt
--index-url https://download.pytorch.org/whl/cpu
torch
--index-url https://pypi.org/simple
grpcio
protobuf
grpcio-reflection
opencv-python-headless
numpy
scipy
gradio
trimesh

